refactor: split api in multiple routes
