+++ package.json
@@ -30,6 +30,7 @@
     "sync:push": "LATEST=$(ls -t .ponder/*_ponder_sync.sql.gz 2>/dev/null | head -n 1); TAG=$(basename \"$LATEST\" | sed 's/_ponder_sync.sql.gz$//'); cp \"$LATEST\" ponder_sync.sql.gz; REPO=$(node -p \"(() => { const repo=require('./package.json').repository; const url=typeof repo==='string'?repo:repo?.url; return new URL(url).pathname.replace(/^\\\\//,'').replace(/\\\\.git$/,''); })()\"); USER=$(gh api user -q .login); gh auth token | oras login ghcr.io -u \"$USER\" --password-stdin; oras push ghcr.io/$REPO/ponder-sync:$TAG ponder_sync.sql.gz:application/gzip; oras tag ghcr.io/$REPO/ponder-sync:$TAG latest; rm -f ponder_sync.sql.gz"
   },
   "dependencies": {
+    "@hono/standard-validator": "^0.2.1",
     "@ponder/utils": "0.2.18",
     "bn.js": "^5.2.3",
     "dotenv": "^16.3.1",
@@ -42,7 +43,8 @@
     "pino": "^8.16.2",
     "pino-pretty": "^13.1.3",
     "ponder": "0.16.6",
-    "viem": "^2.37.11"
+    "viem": "^2.37.11",
+    "zod": "^3.24.2"
   },
   "devDependencies": {
     "@commitlint/cli": "^19.8.1",
