+++ src/api/shared.ts
@@ -1,0 +1,4 @@
+export const jsonDefaultHeaders = {
+  charset: "utf-8",
+  "Content-Type": "application/json",
+} as const;
