+++ src/api/types.ts
@@ -1,0 +1,3 @@
+import { db } from "ponder:api";
+
+export type ApiContext = { db: typeof db };
